Rewrite puppet-sahara from scratch
During the Kilo summit, we talked about the state of the current
puppet-sahara module on stackforge, and especially about the brought up parity with
other modules.

For theses reasons, the rewrite was launched/leaded by Robbie Harwood,
but directly coded on Github (and not in a feature branch in this
repository), for this reason, this patch is non-atomic, and import the
latest frozencemetery/puppet-sahara version¹.

This rewrite/import was approved on puppet-openstack mailing list
(following this thread²).

The «reset» of the stackforge repo was discussed (in order to keep the history
of Robbie module) on #openstack-infra, but it seems we have no simple solution.

¹https://github.com/frozencemetery/puppet-sahara/commit/05af265b15eb3c4e1682ab3fc6dc8b6d750d3e36
²https://groups.google.com/a/puppetlabs.com/forum/#!topic/puppet-openstack/0BETdNvrd70

Co-Authored-By: Robbie Harwood <rharwood@redhat.com>
Change-Id: Ic3dd7547fdfcba275d8a8b90b8ba84bc3ce1d7c0