Fix Beaker CI according recent puppetlabs-rabbitmq change.
The puppetlabs-rabbitmq module introduce a new feature related to the
erlang_cookie parameter change (the database must be wiped if the cookie is
changed), the parameter wipe_db_on_cookie_change is configured to false by
default.

But we're not setting up a rabbitmq cluster and we're not using pacemaker in
beaker CI tests, so we just remove the erlang cookie parameter change.

Change-Id: Id24f65f5a7d7c610353d9ecbf51236c97c7e4cb4
Closes-Bug: #1468444