Creation of sahara::db::sync
In order to standardize the way dbsync are run across our modules,
we create a new class sahara::db::sync.
This class will be included if sync_db is enabled.

By making this transition the sahara::db::sync class
can be returned by the ENC.

A use case would be in an highly available environment, with 3 galera
nodes, include sahara on every node with sync_db set to false
and have the ENC return sahara::db::sync just for one node.

Change-Id: Ibdf6f86e546868b4fcfa0f530287e712934808fd