Add tag to package and service resources
In order to be able to take an action after all the packages of the
module have been installed/updated or all the services have been
started/restarted, we set a 'sahara-packaage' and 'sahara-service' tag
for each package and service of this module.

At the moment, there is a generic openstack tag that is not specific
enough if one wants to take action upon a single module change.

Use case :

If an action needs to be taken after all the packages have been
installed or updated : Package <| tag == 'sahara-package' |> -> X

Change-Id: Ie2ef4a09fbc83a7b97f1d1d5b1d12a3b3a643d34