Change worker defaults to ::os_workers
This patch changes the default worker count from ::processorcount to the
new ::os_workers fact. ::os_workers is based on the number of processors
(currently cpu/4) but is capped at a maximum of 8 worker processors.
This is a much more reasonable default in general and prevents excessive
resource consumption on systems with a large number of CPUs. Also remove
some unnecessary processorcount facts for tests that do not test them.

Change-Id: I96d4ced9955a72a726b7f35b4d5470affc49f4ff