Deprecate the sahara-all service
It's deprecated and is not recommend for usage by upstream as
it has minimal testing.

This adds notes on what should be removed/changed when sahara-all
is to be removed, adds a warning to the sahara::service::all class
and adds a detailed release note on why and how to move away from it.

Change-Id: I7a208813a396f9434ac512ec10e8355aa05d56f9