Configuration: populate the [trustee] section too
Sahara have been (mis)using few values from keystone_authtoken to set the
configuration for trust for a long while. This is not a recommended
practice as those values are for private use of keystonemiddelware.

During the queens release a new configuration section (trustee)
was introduced to collect those parameters:
https://review.openstack.org/#/c/524936/
The existing parameters are still read if [trustee] is not available,
but the behavior is fragile and unsupported, so it is officially deprecated
in Sahara as well.

The trustee section is populated from the same values used for
keyston_authtoken, so the same class is re-used in order to write
the content of the section.

Closes-Bug: #1763361
Change-Id: I8c63c32d995cbe4140680d826869c25d855e77d1